# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShareEd Backend API",
    "version": "1.0.0",
    "description": "Complete API documentation for ShareEd Backend. Built using Express and Prisma."
  },
  "servers": [
    {
      "url": "https://share-ed-backend-6jer.onrender.com/api/v1",
      "description": "deploy server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Supabase Access Token (JWT)"
      }
    },
...
  },
  "paths": {
...
    "/likes/{postId}": {
      "get": {
        "summary": "Like Status",
        "description": "Checks if the current user has liked the post.",
        "tags": [
          "❤️ Likes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Like status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "liked": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "post": {
        "summary": "Toggle Like/Unlike",
        "description": "Like the post if not liked yet, or unlike if already liked.",
        "tags": [
          "❤️ Likes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Toggled status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "liked": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  }
}